refactor: use samplesheet inputs to determine tool dependency
Keeping the current paradigm of "supply inputs you have, pipeline
generates the rest" while adding the addendum of skipping unnecessary
tools. This required moving all of the intermediate channels into the
global scope and a less naive way of walking through the graph of tool
dependencies.

E.g.:
1. You have hets, coverage, snvs, and svs and you want to run everything
you can starting with these inputs.
2. You have bams, coverage and svs, and you want to run everything you
can starting with these inputs.
3. You have bams, coverage, and svs, and you only want to run JaBbA and nothing else.

Requirements:
- user can give subset of inputs and pipeline will run all tools it can
with those inputs, by stepping through dependency tree
- user can specify tools to skip via a skip_tools parameter

Given:
- pipeline tool dependency tree
- user inputs
- user tools

Strategy:
In all cases, the existing outputs instantiation will need to be moved to
the global scope. I.e first populate a channel with the existing outputs
outside of the process execution block, then if that block is run, mix its
outputs to the global channel. This way, even if the block is not run, the
outputs will still be available to the next block.

Create a map of tools to their required inputs, then just use brute
force iteration/recursion. Maintain a list of available inputs, insert
the user inputs first, then iterate through the tools and check if the
inputs of each tool are satisfied by the available inputs and the
outputs of the tools are not already in the available inputs. If so, add
the tool to the list of tools to be run and add its outputs to the
available inputs. Then recurse with the new available inputs until no
new tools can be added. Ignore any tools which are in the skip_tools
list.